requirements.txt
python-telegram-bot==20.7
yt-dlp==2024.11.18
python-dotenv==1.0.0
pytest==7.4.3
